requirements/base.txt
Django==1.5.1
MySQL-python==1.2.4
Pillow==2.1.0
South==0.8.1
Unidecode==0.04.13
Unipath==1.0
argparse==1.2.1
boto==2.9.9
django-appconf==0.6
django-braces==1.1.0
django-categories==1.2.1
django-ckeditor==4.0.2
django-debug-toolbar==0.9.4
django-model-utils==1.4.0
django-mptt==0.5.5
django-pipeline==1.3.13
django-social-auth==0.7.23
django-staticfiles==1.2.1
django-storages==1.1.8
django-tinymce==1.5.1
django-widget-tweaks==1.3
django-wysiwyg==0.5.1
djangorestframework==2.3.6
futures==2.1.4
httplib2==0.8
oauth2==1.5.211
python-openid==2.2.5
python-slugify==0.0.6
six==1.4.1
sorl-thumbnail==11.12
wsgiref==0.1.2
yuicompressor==2.4.7
